smaller

+++ Q2/SDS.py
@@ -154,7 +154,7 @@ class SDS:
         # predict the noise residual with unet, NO grad!
         with torch.no_grad():
             ### YOUR CODE HERE ###
-            epsilon = torch.randn(B, 4, self.H, self.W).cuda()
+            epsilon = torch.randn(B, 4, self.H, self.W, dtype = torch.float16).cuda()
             eps_hat = self.unet(epsilon, t, text_embeddings)
             noise_residual = torch.linalg.norm(eps_hat - epsilon)
